api/requirements.txt
fastapi==0.115.1
uvicorn==0.31.1
openai==1.57.4
anthropic==0.40.0
psycopg2-binary==2.9.9
numpy==1.26.1
requests==2.31.0
pandas==2.1.1
